EzikYırtıkKargoAyıklama/requirements.txt
tensorflow==2.15.0
flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
requests>=2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
gdown>=4.4.0
tqdm>=4.65.0
matplotlib==3.8.2